feat!: add skeleton `AnalysisEpic` to demonstrate PODIO usage
This begins the new `Analysis`-derived class, `AnalysisEpic`, for reading full simulation output from the EPIC software stack. It requires PODIO libraries, and the EPIC data model (EDM4hep+eicd), all of which are available in `jug_xl`; consequently, merging this PR will *require* users to run in `jug_xl`, or to build PODIO + the data model on their own.

This PR is for basic functionality only; a new PR should be made for the first "full" version of `AnalysisEpic`, following a similar structure as the other `Analysis`-derived classes.

**TODO**:
- [ ] update documentation, informing users they need to use `jug_xl`
- [ ] test this on more "complete" reconstructed data, since:
```
// currently testing with files produced from benchmarks:
//   repo:        physics_benchmarks, from https://eicweb.phy.anl.gov/EIC/benchmarks/physics_benchmarks
//   CI stage:    finish
//   CI job:      summary
//   CI artifact: results/dis/10on100/minQ2=1/rec-dis_10x100_minQ2=1.root
```