feat: control what is included in the output
Add `includeOutputSet` settings in `Analysis`, to control what to
include in the output. This will allow us to turn on/off large sets of
plots or tree branches, along with certain parts of the `Analysis`
subclass event loops. For example, we can now implement a dihadron
analysis which will only run if a macro tells it to.